requirements.txt
google-generativeai
flask
instaloader